Create LongestWord
This challenge requires you to return the longest word in a sentence. Our goal is therefore to somehow loop through the string (what is a string?) and compare the lengths of the words. For example, if the string were "hello from coderbyte" the program should return the word "coderbyte" because it has the longest length. To solve this we need to:

1. Strip away any punctuation because the challenge tells us to ignore these symbols, so the string "hello$%()" is actually just "hello"
2. Separate the sentence into a list of words in order to easily retrieve words and their respective lengths
3. Then loop through this list and compare the words to find the one with the longest length 

We will be using regular expression functions in order to strip away punctuation and turn the string into a list of words. Regular expressions are patterns we define for the program to find in some text.

Below is some code that performs each of these steps: